refactor(atoms): merge SfQuantitySelector
