--- packages/vue/src/components/atoms/SfQuantitySelector/SfQuantitySelector.html
@@ -1,8 +1,0 @@
-<SfInput
-  type="number"
-  :value="qty"
-  :min="min"
-  :max="stock"
-  class="sf-quantity-selector"
-  @input="$emit('input', parseInt($event,10))"
-/>
